+++ Expanded Cheat Menu/Changelog.md
@@ -6,6 +6,8 @@
 ## v8.6 (Unreleased)
 ### Added:
 - Rockets>Max Export Amount: Change how many rares per rocket you can export.
+- Rockets>Launch Fuel Per Rocket: Change how much fuel rockets need to launch.
+- Rockets>Rockets Ignore Fuel: Rockets don't need fuel to launch.
 
 ### Changed:
 - Cheats>Research>Add Points: Added option to reset sponsor points to default.